+++ tests/config/test_config.py
@@ -8,7 +8,7 @@ import pytest
 
 import libs.global_value as g
 from cls.config import SubCommand
-from libs import configuration
+from libs.bootstrap import configuration
 
 
 def test_empty_config(monkeypatch):
