+++ tests/parser/test_parser.py
@@ -8,7 +8,7 @@ import pytest
 
 import libs.global_value as g
 from integrations import factory
-from libs import configuration
+from libs.bootstrap import configuration
 from libs.domain.command import CommandParser
 from libs.utils import dictutil
 from tests.parser import param_data
