+++ tests/parser/test_score.py
@@ -8,7 +8,7 @@ import pytest
 
 import libs.global_value as g
 from integrations import factory
-from libs import configuration
+from libs.bootstrap import configuration
 from libs.domain.score import GameResult
 from libs.utils import validator
 from tests.parser import param_data
