+++ tests/test.py
@@ -12,14 +12,14 @@ from pprint import pprint
 
 import libs.global_value as g
 from integrations import factory
-from libs import configuration
+from libs.bootstrap import configuration
+from libs.bootstrap.configuration import arg_parser
 from libs.commands.graph import personal as graph_personal
 from libs.commands.graph import rating as graph_rating
 from libs.commands.graph import summary as graph_summary
 from libs.commands.ranking import ranking
 from libs.commands.report import stats_list, stats_report
 from libs.commands.results import summary as results_summary
-from libs.configuration import arg_parser
 from libs.domain.command import CommandParser
 from libs.functions import compose
 from libs.utils import dictutil
